# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hostreamly Media API",
    "version": "2.0.0",
    "description": "API completa para la plataforma Hostreamly Media con funcionalidades empresariales avanzadas. Esta API proporciona endpoints para gestión de videos, streaming en vivo, analíticas avanzadas, webhooks personalizados, formularios de contacto, enlaces de descarga controlados y mucho más.",
    "termsOfService": "https://hostreamly.com/terms",
    "contact": {
      "name": "Hostreamly Support",
      "email": "support@hostreamly.com",
      "url": "https://hostreamly.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.hostreamly.com/api",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "WhiteLabelConfig": {
        "type": "object",
        "properties": {
          "brandName": {
            "type": "string"
          },
          "logo": {
            "type": "string",
            "format": "uri"
          },
          "favicon": {
            "type": "string",
            "format": "uri"
          },
          "primaryColor": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$"
          },
          "secondaryColor": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$"
          },
          "customDomain": {
            "type": "string"
          },
          "customCSS": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "CustomDomain": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "verified": {
            "type": "boolean"
          },
          "verificationToken": {
            "type": "string"
          },
          "dnsRecords": {
            "type": "object",
            "properties": {
              "cname": {
                "type": "string"
              },
              "txt": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/white-label/config": {
      "get": {
        "summary": "Obtener configuración de marca blanca",
        "tags": ["Marca Blanca"],
        "responses": {
          "200": {
            "description": "Configuración actual de marca blanca",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteLabelConfig"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Actualizar configuración de marca blanca",
        "tags": ["Marca Blanca"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhiteLabelConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuración actualizada exitosamente"
          }
        }
      }
    },
    "/white-label/domains": {
      "get": {
        "summary": "Obtener dominios personalizados",
        "tags": ["Marca Blanca"],
        "responses": {
          "200": {
            "description": "Lista de dominios personalizados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomDomain"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Configurar nuevo dominio personalizado",
        "tags": ["Marca Blanca"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "domain": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dominio configurado exitosamente"
          }
        }
      }
    },
    "/white-label/domains/{domain}/verify": {
      "post": {
        "summary": "Verificar dominio personalizado",
        "tags": ["Marca Blanca"],
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dominio verificado exitosamente"
          }
        }
      }
    },
...
  },
...
}